fix  circular dependency among user->company-shared
