+++ backend/src/models/internalTransaction.ts
@@ -2,7 +2,7 @@ import { getModelForClass, prop, Ref } from '@typegoose/typegoose';
 
 import { Base } from './base';
 import { Contract } from './contract';
-import { EditionHistoryItem } from './shared';
+import { EditionHistoryItem } from './shared/editionHistoryItem';
 import { Team } from './team';
 import { User } from './user';
