+++ database/1_6_0.mongodb
@@ -144,3 +144,9 @@ db.companies.find().forEach((company) => {
   };
   db.companies.replaceOne({ _id: company._id }, company);
 });
+
+db.users.find().forEach((user) => {
+  const company = db.companies.findOne();
+  user.company = company;
+  db.users.replaceOne({ _id: user._id }, user);
+});
